POC: Running async jobs via Artemis
This is a POC around running async jobs with Artemis.

Most of the classes of interest are in the following package:
    org.candlepin.async

AsyncResource is a good starting point when looking to see how
this works.

DISCLAIMER
==========
This branch is a very crude implementation of sending job messages
that trigger async tasks via a ThreadPoolExecutor. Do not expect the
tests to run. There are a lot of FIXME/TODOs in this code. The intent
is to get something working and allow us to better estimate what would
be involved in the full implementation. A lot of this can be reused,
but will require some cleanup and testing.